Doubling down on the BackupStrategy. Expecting to move this into both 'ac' as well as the GUI.
