+++ app/cli-acx/src/main/java/io/github/applecommander/acx/command/CreateDiskCommand.java
@@ -27,7 +27,7 @@ import com.webcodepro.applecommander.storage.os.pascal.PascalFormatDisk;
 import com.webcodepro.applecommander.storage.os.prodos.ProdosFormatDisk;
 import io.github.applecommander.acx.OrderType;
 import io.github.applecommander.acx.SystemType;
-import io.github.applecommander.acx.base.BackupStrategy;
+import org.applecommander.util.BackupStrategy;
 import io.github.applecommander.acx.base.ReusableCommandOptions;
 import io.github.applecommander.acx.converter.DataSizeConverter;
 import org.applecommander.device.*;
