+++ lib/ac-api/src/main/java/org/applecommander/util/BackupStrategy.java
@@ -17,7 +17,7 @@
  * with this program; if not, write to the Free Software Foundation, Inc.,
  * 59 Temple Place, Suite 330, Boston, MA 02111-1307 USA
  */
-package io.github.applecommander.acx.base;
+package org.applecommander.util;
 
 import java.io.IOException;
 import java.nio.file.Files;
@@ -26,12 +26,45 @@ import java.nio.file.Path;
 import static java.nio.file.StandardCopyOption.REPLACE_EXISTING;
 
 /// Provides a simple but automated backup strategy to the `acx` tool.
+///
+/// General usage pattern is as follows:
+/// ```java
+/// // Only save if there are changes.
+/// if (disk.getSource().hasChanged()) {
+///     backupStrategy.backup(disk.getFilename());
+///     disk.save();
+/// }
+/// ```
+///
+/// To create a `BackupStrategy` in a consistent manner, use the #create
+/// method:
+/// ```'java
+/// BackupStrategy backupStrategy = BackupStrategy.create("bak");
+/// ```
+///
+/// If using in a context where a backup is no expected (as in a new file, or
+/// a "save as" situation), simply use `BackupStrategy.none()`.
+///
+/// The coding is as follows:
+/// * `null` or blank (`""`): Returns an ineffective strategy. (No branches in code.)
+/// * `bak`: Returns a backup strategy that renames the existing file from `name.dsk`
+///   to `name.dsk.bak`, overwriting any existing files.
+/// * The default strategy is to assume the value is a directory, _which will be created_,
+///   and make a copy of the file into this directory.
+///
+/// @see #create(String)
 /// @see #none()
-/// @see #fileExtension(String) 
-/// @see #directory(String)
-/// @see ReadWriteDiskCommandOptions
 public interface BackupStrategy {
         void backup(String filename) throws IOException;
+
+        static BackupStrategy create(String backupCode) {
+            if (backupCode == null) backupCode = "";
+            return switch (backupCode) {
+                case "" -> BackupStrategy.none();
+                case "bak" -> BackupStrategy.fileExtension("bak");
+                default -> BackupStrategy.directory(backupCode);
+            };
+        }
 
         static BackupStrategy none() {
             return filename -> {
